backend/requirements.txt
# Core Framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.11.7
pydantic-settings>=2.1.0

# FastMCP - Modern MCP Implementation
fastmcp==2.10.1

# MongoDB Database
motor==3.3.2
pymongo==4.6.0
beanie==1.23.6

# Redis for Caching
redis==5.0.1
aioredis==2.0.1

# HTTP Client for Ollama
httpx>=0.28.1
aiohttp==3.9.1

# GraphQL
graphql-core==3.2.3
strawberry-graphql==0.215.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.9

# Data Validation & Serialization
marshmallow==3.20.1
jsonschema==4.20.0

# Background Tasks
celery==5.3.4
flower==2.0.1

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.39.2
prometheus-client==0.19.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Environment & Configuration
python-dotenv>=1.1.0
click==8.1.7

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Text Processing
nltk==3.8.1
# spacy==3.7.2

# Additional Utilities
rich>=13.9.4
tenacity==8.2.3

# Enhanced LLM Tracking
psutil==5.9.6

# Event Loop Patching
nest-asyncio>=1.5.6

# Event Loop Patching
nest_asyncio 